Refactor messages, split files, add more message types from APK
